+++ CHANGELOG.md
@@ -7,6 +7,14 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ## [Unreleased]
 
+## [v1.1.1] - 2024-08-22
+
+### Added
+- Add new `TestRequest` class extending `Request` for tests
+
+### Updated
+- Enhance tests with more pre-built tests
+
 ## [v1.1.0] - 2024-08-20
 
 ### Added